Extract shared runner, eliminate wire format conversion, add loading spinner
- Extract test execution logic from worker.py into runner.py, reducing
  worker.py by ~270 lines of duplicated code
- Create playground.py as a thin harness that delegates to runner.run_test
- Move RunTestResultWire and AssertionWire to tryke_types so the WASM
  crate can accept the flat runner format directly, eliminating the
  Python-side _to_wire conversion entirely
- Extract inline example Python code from constants.ts into standalone
  .py files under playground/examples/ so ruff can lint/format them
- Add playground/examples to ruff src config with appropriate per-file
  ignores for example code
- Show a loading spinner in the output panel while Pyodide loads
- Pass pyodideReady through to Editor for conditional rendering
