+++ crates/tryke_runner/src/protocol.rs
@@ -121,65 +121,4 @@ pub struct RunDoctestParams {
     pub object_path: String,
 }
 
-#[derive(Debug, Deserialize)]
-#[serde(tag = "outcome", rename_all = "snake_case")]
-pub enum RunTestResultWire {
-    Passed {
-        duration_ms: u64,
-        stdout: String,
-        stderr: String,
-    },
-    Failed {
-        duration_ms: u64,
-        message: String,
-        #[serde(default)]
-        traceback: Option<String>,
-        #[serde(default)]
-        assertions: Vec<AssertionWire>,
-        #[serde(default)]
-        executed_lines: Vec<u32>,
-        stdout: String,
-        stderr: String,
-    },
-    Skipped {
-        duration_ms: u64,
-        #[serde(default)]
-        reason: Option<String>,
-        stdout: String,
-        stderr: String,
-    },
-    #[serde(rename = "xfailed")]
-    XFailed {
-        duration_ms: u64,
-        #[serde(default)]
-        reason: Option<String>,
-        stdout: String,
-        stderr: String,
-    },
-    #[serde(rename = "xpassed")]
-    XPassed {
-        duration_ms: u64,
-        stdout: String,
-        stderr: String,
-    },
-    Todo {
-        duration_ms: u64,
-        #[serde(default)]
-        description: Option<String>,
-        stdout: String,
-        stderr: String,
-    },
-}
-
-#[derive(Debug, Deserialize)]
-pub struct AssertionWire {
-    pub expression: String,
-    pub expected: String,
-    pub received: String,
-    #[serde(default)]
-    pub line: u32,
-    #[serde(default)]
-    pub column: Option<u32>,
-    #[serde(default)]
-    pub file: Option<String>,
-}
+pub use tryke_types::{AssertionWire, RunTestResultWire};
