+++ playground/examples/helpers.py
@@ -1,0 +1,6 @@
+def double(n):
+    return n * 2
+
+
+def greet(name):
+    return f"hello, {name}"
