+++ playground/examples/mathlib.py
@@ -1,0 +1,16 @@
+def add(a, b):
+    return a + b
+
+
+def multiply(a, b):
+    return a * b
+
+
+def divide(a, b):
+    if b == 0:
+        raise ValueError("division by zero")
+    return a / b
+
+
+def clamp(value, low, high):
+    return max(low, min(value, high))
