+++ playground/src/Editor/Editor.tsx
@@ -35,6 +35,29 @@ function Section({
   );
 }
 
+function LoadingSpinner() {
+  return (
+    <div className="h-full flex flex-col items-center justify-center gap-3 text-text-dim">
+      <svg className="animate-spin h-6 w-6" viewBox="0 0 24 24" fill="none">
+        <circle
+          className="opacity-25"
+          cx="12"
+          cy="12"
+          r="10"
+          stroke="currentColor"
+          strokeWidth="3"
+        />
+        <path
+          className="opacity-75"
+          fill="currentColor"
+          d="M4 12a8 8 0 018-8V0C5.373 0 0 5.373 0 12h4z"
+        />
+      </svg>
+      <span className="text-xs">Loading Python...</span>
+    </div>
+  );
+}
+
 interface WasmModule {
   discover: (source: string, filename: string) => DiscoveredFile;
   discover_multi: (files_json: string) => {
@@ -55,6 +78,7 @@ interface Props {
   terminalOutput: string;
   runStatus: RunStatus;
   wasm: WasmModule | null;
+  pyodideReady: boolean;
 }
 
 export function Editor({
@@ -67,6 +91,7 @@ export function Editor({
   terminalOutput,
   runStatus,
   wasm,
+  pyodideReady,
 }: Props) {
   const activeFile = files[activeFileIndex]!;
   const deferredSource = useDeferredValue(activeFile.source);
@@ -109,6 +134,7 @@ export function Editor({
   }, [wasm, discovery, reporter]);
 
   const displayOutput = runStatus !== "idle" ? terminalOutput : collectOutput;
+  const showLoading = !pyodideReady && !displayOutput;
 
   return (
     <PanelGroup direction="horizontal" className="h-full">
@@ -132,7 +158,11 @@ export function Editor({
               <div className="h-full flex flex-col">
                 <div className="h-1/2 min-h-0 border-b border-border">
                   <Section title="Output" fill>
-                    <TerminalOutput content={displayOutput} />
+                    {showLoading ? (
+                      <LoadingSpinner />
+                    ) : (
+                      <TerminalOutput content={displayOutput} />
+                    )}
                   </Section>
                 </div>
                 <div className="h-1/2 min-h-0 flex flex-col">
@@ -170,9 +200,12 @@ export function Editor({
                 {secondaryTool === "fixture-graph" && (
                   <FixtureGraphView hooks={discovery?.parsed.hooks ?? []} />
                 )}
-                {secondaryTool === "output" && (
-                  <TerminalOutput content={displayOutput} />
-                )}
+                {secondaryTool === "output" &&
+                  (showLoading ? (
+                    <LoadingSpinner />
+                  ) : (
+                    <TerminalOutput content={displayOutput} />
+                  ))}
               </>
             )}
           </div>
