+++ python/tryke/worker.py
@@ -20,8 +20,8 @@ Supported methods:
 - `run_doctest {module, object_path}` → tagged outcome dict
 
 The wire format mirrors `crates/tryke_runner/src/protocol.rs`. The
-TypedDicts below (`_AssertionWire`, `_PassedResult`, …) are the result
-shapes the runner decodes. The hook-related request shapes are handled
+result TypedDicts live in :mod:`tryke.runner` and are shared with the
+playground. The hook-related request shapes are handled
 at the dispatch boundary in `_register_hooks` — the Rust side
 statically discovered the fixture metadata with Ruff, so the worker
 just trusts the incoming list and does not re-parse source.
@@ -49,11 +49,9 @@ just trusts the incoming list and does not re-parse source.
 
 from __future__ import annotations
 
-import asyncio
 import contextlib
 import doctest
 import importlib
-import inspect
 import io
 import json
 import logging
@@ -61,23 +59,11 @@ import os
 import sys
 import time
 import traceback
-import unittest
-from pathlib import Path
-from typing import TYPE_CHECKING, Literal, NotRequired, TypedDict
+from typing import TYPE_CHECKING
 
 import tryke_guard
-from tryke.expect import (
-    CaseArgs,
-    CasesMarked,
-    ExpectationError,
-    SoftContext,
-    SoftFailure,
-    _set_soft_context,
-    _SkipMarked,
-    _TodoMarked,
-    _XfailMarked,
-)
 from tryke.hooks import HookExecutor, _fixture_per
+from tryke.runner import TestResult, failed, passed, run_test
 
 # Flip `tryke_guard.__TRYKE_TESTING__` on for this worker process. User
 # modules imported later via `_get_module` do
@@ -91,266 +77,16 @@ from tryke.hooks import HookExecutor, _fixture_per
 tryke_guard.__TRYKE_TESTING__ = True
 
 if TYPE_CHECKING:
-    from collections.abc import Generator
     from types import ModuleType
     from typing import TextIO
 
-_TRYKE_PKG = str(Path(__file__).resolve().parent)
-
-# Worker-side logger. Debug/trace output is useful when auditing module
-# import and dispatch; enable with `TRYKE_WORKER_LOG=DEBUG` (or `TRACE`).
-# Messages go to stderr so they don't corrupt the JSON-RPC channel on
-# stdout. Off by default.
 _log = logging.getLogger("tryke.worker")
 
-
-# -- Wire-format TypedDicts (mirror crates/tryke_runner/src/protocol.rs) ------
-
-
-class _AssertionWire(TypedDict):
-    expression: str
-    expected: str
-    received: str
-    line: NotRequired[int]
-    column: NotRequired[int]
-    file: NotRequired[str]
-
-
-class _PassedResult(TypedDict):
-    outcome: Literal["passed"]
-    duration_ms: int
-    stdout: str
-    stderr: str
-
-
-class _FailedResult(TypedDict):
-    outcome: Literal["failed"]
-    duration_ms: int
-    message: str
-    traceback: str | None
-    assertions: list[_AssertionWire]
-    executed_lines: list[int]
-    stdout: str
-    stderr: str
-
-
-class _SkippedResult(TypedDict):
-    outcome: Literal["skipped"]
-    duration_ms: int
-    reason: str | None
-    stdout: str
-    stderr: str
-
-
-class _XFailedResult(TypedDict):
-    outcome: Literal["xfailed"]
-    duration_ms: int
-    reason: str | None
-    stdout: str
-    stderr: str
-
-
-class _XPassedResult(TypedDict):
-    outcome: Literal["xpassed"]
-    duration_ms: int
-    stdout: str
-    stderr: str
-
-
-class _TodoResult(TypedDict):
-    outcome: Literal["todo"]
-    duration_ms: int
-    description: str | None
-    stdout: str
-    stderr: str
-
-
-type _TestResult = (
-    _PassedResult
-    | _FailedResult
-    | _SkippedResult
-    | _XFailedResult
-    | _XPassedResult
-    | _TodoResult
-)
-
-type _DispatchResult = _TestResult | str | None
+type _DispatchResult = TestResult | str | None
 
 
 class _InvalidParamsError(Exception):
     """Missing or invalid JSON-RPC method parameter."""
-
-
-def _passed(
-    duration_ms: int,
-    stdout: str,
-    stderr: str,
-) -> _PassedResult:
-    return {
-        "outcome": "passed",
-        "duration_ms": duration_ms,
-        "stdout": stdout,
-        "stderr": stderr,
-    }
-
-
-def _failed(  # noqa: PLR0913
-    duration_ms: int,
-    message: str,
-    tb: str | None,
-    assertions: list[_AssertionWire],
-    stdout: str,
-    stderr: str,
-    *,
-    executed_lines: list[int] | None = None,
-) -> _FailedResult:
-    return {
-        "outcome": "failed",
-        "duration_ms": duration_ms,
-        "message": message,
-        "traceback": tb,
-        "assertions": assertions,
-        "executed_lines": executed_lines if executed_lines is not None else [],
-        "stdout": stdout,
-        "stderr": stderr,
-    }
-
-
-def _skipped(
-    duration_ms: int,
-    reason: str | None,
-    stdout: str,
-    stderr: str,
-) -> _SkippedResult:
-    return {
-        "outcome": "skipped",
-        "duration_ms": duration_ms,
-        "reason": reason,
-        "stdout": stdout,
-        "stderr": stderr,
-    }
-
-
-def _xfailed(
-    duration_ms: int,
-    reason: str | None,
-    stdout: str,
-    stderr: str,
-) -> _XFailedResult:
-    return {
-        "outcome": "xfailed",
-        "duration_ms": duration_ms,
-        "reason": reason,
-        "stdout": stdout,
-        "stderr": stderr,
-    }
-
-
-def _xpassed(
-    duration_ms: int,
-    stdout: str,
-    stderr: str,
-) -> _XPassedResult:
-    return {
-        "outcome": "xpassed",
-        "duration_ms": duration_ms,
-        "stdout": stdout,
-        "stderr": stderr,
-    }
-
-
-def _todo(
-    duration_ms: int,
-    description: str | None,
-    stdout: str,
-    stderr: str,
-) -> _TodoResult:
-    return {
-        "outcome": "todo",
-        "duration_ms": duration_ms,
-        "description": description,
-        "stdout": stdout,
-        "stderr": stderr,
-    }
-
-
-def _is_user_frame(frame: traceback.FrameSummary) -> bool:
-    return not str(
-        Path(frame.filename).resolve(),
-    ).startswith(_TRYKE_PKG)
-
-
-def _make_assertion_wire(
-    *,
-    expression: str,
-    expected: str,
-    received: str,
-    frame: traceback.FrameSummary | None = None,
-) -> _AssertionWire:
-    wire: _AssertionWire = {
-        "expression": expression,
-        "expected": expected,
-        "received": received,
-    }
-    if frame is not None:
-        if frame.lineno is not None:
-            wire["line"] = frame.lineno
-        if frame.colno is not None:
-            wire["column"] = frame.colno
-        wire["file"] = frame.filename
-    return wire
-
-
-def _extract_soft_failures(
-    failures: list[SoftFailure],
-) -> list[_AssertionWire]:
-    return [
-        _make_assertion_wire(
-            expression=(frame.line or "").strip() if frame else "",
-            expected=err.expected,
-            received=err.received,
-            frame=frame,
-        )
-        for err, frame in failures
-    ]
-
-
-def _extract_single(exc: ExpectationError) -> _AssertionWire:
-    tb = sys.exc_info()[2]
-    frames = traceback.extract_tb(tb)
-    for frame in reversed(frames):
-        if _is_user_frame(frame):
-            return _make_assertion_wire(
-                expression=(frame.line or "").strip(),
-                expected=exc.expected,
-                received=exc.received,
-                frame=frame,
-            )
-    return _make_assertion_wire(
-        expression="",
-        expected=exc.expected,
-        received=exc.received,
-    )
-
-
-def _extract_assertions(
-    exc: AssertionError,
-) -> list[_AssertionWire]:
-    if not isinstance(exc, ExpectationError):
-        return []
-    tb = sys.exc_info()[2]
-    frames = traceback.extract_tb(tb)
-    for frame in reversed(frames):
-        if _is_user_frame(frame):
-            return [
-                _make_assertion_wire(
-                    expression=(frame.line or "").strip(),
-                    expected=exc.expected,
-                    received=exc.received,
-                    frame=frame,
-                )
-            ]
-    return []
 
 
 class Worker:
@@ -569,18 +305,7 @@ class Worker:
         self._executors[module_name] = executor
         return executor
 
-    @contextlib.contextmanager
-    def _soft_assertion_context(
-        self,
-    ) -> Generator[SoftContext, None, None]:
-        ctx = SoftContext()
-        _set_soft_context(ctx)
-        try:
-            yield ctx
-        finally:
-            _set_soft_context(None)
-
-    def _run_test(  # noqa: C901, PLR0911, PLR0912, PLR0915
+    def _run_test(
         self,
         module_name: str,
         function_name: str,
@@ -588,12 +313,12 @@ class Worker:
         xfail: str | None = None,
         groups: list[str] | None = None,
         case_label: str | None = None,
-    ) -> _TestResult:
+    ) -> TestResult:
         try:
             mod = self._get_module(module_name)
             fn = getattr(mod, function_name)
         except Exception as exc:  # noqa: BLE001
-            return _failed(
+            return failed(
                 0,
                 f"{type(exc).__name__}: {exc}",
                 traceback.format_exc(),
@@ -602,201 +327,28 @@ class Worker:
                 "",
             )
 
-        case_args: tuple[object, ...] = ()
-        case_kwargs: CaseArgs | None = None
-        if case_label is not None:
-            if not isinstance(fn, CasesMarked):
-                return _failed(
-                    0,
-                    (
-                        f"{function_name} was dispatched with case_label="
-                        f"{case_label!r} but has no __tryke_cases__ attribute"
-                    ),
-                    "",
-                    [],
-                    "",
-                    "",
-                )
-            cases = fn.__tryke_cases__
-            entry = next((e for e in cases if e.label == case_label), None)
-            if entry is None:
-                known = sorted(e.label for e in cases)
-                return _failed(
-                    0,
-                    (
-                        f"{function_name} has no case labeled "
-                        f"{case_label!r}; known cases: {known}"
-                    ),
-                    "",
-                    [],
-                    "",
-                    "",
-                )
-            case_args = entry.args
-            case_kwargs = entry.kwargs
-
-            # Per-case modifiers (runtime fallback for non-literal values
-            # that static discovery could not extract).
-            if entry.skip is not None:
-                return _skipped(0, entry.skip, "", "")
-            if entry.todo is not None:
-                return _todo(0, entry.todo, "", "")
-            if entry.xfail is not None:
-                xfail = entry.xfail
-
-        # Runtime skip/todo (handles skip_if resolved at import time)
-        if isinstance(fn, _SkipMarked):
-            return _skipped(0, fn.__tryke_skip__, "", "")
-
-        if isinstance(fn, _TodoMarked):
-            return _todo(0, fn.__tryke_todo__, "", "")
-
-        is_xfail = xfail is not None or isinstance(fn, _XfailMarked)
-        xfail_reason = (
-            xfail
-            if xfail is not None
-            else (fn.__tryke_xfail__ if isinstance(fn, _XfailMarked) else None)
+        return run_test(
+            fn,
+            executor=self._get_executor(module_name),
+            xfail=xfail,
+            groups=groups,
+            case_label=case_label,
         )
-
-        stdout_buf = io.StringIO()
-        stderr_buf = io.StringIO()
-        start = time.monotonic()
-
-        with self._soft_assertion_context() as ctx:
-            try:
-                with (
-                    contextlib.redirect_stdout(stdout_buf),
-                    contextlib.redirect_stderr(stderr_buf),
-                ):
-                    executor = self._get_executor(module_name)
-                    if executor is not None:
-                        executor.run_test(
-                            fn,
-                            groups=groups or [],
-                            case_args=case_args,
-                            case_kwargs=case_kwargs,
-                        )
-                    elif inspect.iscoroutinefunction(fn):
-                        asyncio.run(fn(*case_args, **(case_kwargs or {})))
-                    else:
-                        fn(*case_args, **(case_kwargs or {}))
-
-                ms = int((time.monotonic() - start) * 1000)
-                out = stdout_buf.getvalue()
-                err = stderr_buf.getvalue()
-
-                if ctx.failures:
-                    if is_xfail:
-                        return _xfailed(
-                            ms,
-                            xfail_reason,
-                            out,
-                            err,
-                        )
-                    return _failed(
-                        ms,
-                        "assertion failed",
-                        "",
-                        _extract_soft_failures(ctx.failures),
-                        out,
-                        err,
-                        executed_lines=list(ctx.executed_lines),
-                    )
-                if is_xfail:
-                    return _xpassed(ms, out, err)
-                return _passed(ms, out, err)
-
-            except unittest.SkipTest as exc:
-                ms = int((time.monotonic() - start) * 1000)
-                return _skipped(
-                    ms,
-                    str(exc),
-                    stdout_buf.getvalue(),
-                    stderr_buf.getvalue(),
-                )
-
-            except ExpectationError as exc:
-                ms = int((time.monotonic() - start) * 1000)
-                out = stdout_buf.getvalue()
-                err = stderr_buf.getvalue()
-                if is_xfail:
-                    return _xfailed(
-                        ms,
-                        xfail_reason,
-                        out,
-                        err,
-                    )
-                assertions = _extract_soft_failures(
-                    ctx.failures,
-                )
-                assertions.append(_extract_single(exc))
-                return _failed(
-                    ms,
-                    str(exc) or "assertion failed",
-                    traceback.format_exc(),
-                    assertions,
-                    out,
-                    err,
-                    executed_lines=list(ctx.executed_lines),
-                )
-
-            except AssertionError as exc:
-                ms = int((time.monotonic() - start) * 1000)
-                out = stdout_buf.getvalue()
-                err = stderr_buf.getvalue()
-                if is_xfail:
-                    return _xfailed(
-                        ms,
-                        xfail_reason,
-                        out,
-                        err,
-                    )
-                return _failed(
-                    ms,
-                    str(exc) or "assertion failed",
-                    traceback.format_exc(),
-                    _extract_assertions(exc),
-                    out,
-                    err,
-                    executed_lines=list(ctx.executed_lines),
-                )
-
-            except Exception as exc:  # noqa: BLE001
-                ms = int((time.monotonic() - start) * 1000)
-                out = stdout_buf.getvalue()
-                err = stderr_buf.getvalue()
-                if is_xfail:
-                    return _xfailed(
-                        ms,
-                        xfail_reason,
-                        out,
-                        err,
-                    )
-                return _failed(
-                    ms,
-                    f"{type(exc).__name__}: {exc}",
-                    traceback.format_exc(),
-                    [],
-                    out,
-                    err,
-                    executed_lines=list(ctx.executed_lines),
-                )
 
     def _run_doctest(
         self,
         module_name: str,
         object_path: str,
-    ) -> _TestResult:
+    ) -> TestResult:
         try:
             mod = self._get_module(module_name)
 
-            # Resolve the target object whose docstring we want to test.
             obj = mod
             if object_path:
                 for attr in object_path.split("."):
                     obj = getattr(obj, attr)
         except Exception as exc:  # noqa: BLE001
-            return _failed(
+            return failed(
                 0,
                 f"{type(exc).__name__}: {exc}",
                 traceback.format_exc(),
@@ -842,7 +394,7 @@ class Worker:
         err = stderr_buf.getvalue()
 
         if summary.failed > 0:
-            return _failed(
+            return failed(
                 ms,
                 output_buf.getvalue(),
                 None,
@@ -850,7 +402,7 @@ class Worker:
                 out,
                 err,
             )
-        return _passed(ms, out, err)
+        return passed(ms, out, err)
 
 
 def _configure_logging_from_env() -> None:
